move getAvailablePort to separate util
